requirements.txt
win_unicode_console
-i http://mirrors.aliyun.com/pypi/simple/ --trusted-host mirrors.aliyun.com
django
django-suit
django-import-export
django-model-utils
django_extensions
django-environ
django-secure
django-celery
django-mptt

django-suit-ckeditor
django-extra-fields

django-debug-toolbar


# django rest framework
djangorestframework-word-filter
djangorestframework-recursive
djangorestframework-filters
django-rest-multiple-models
djangorestframework-oauth
djangorestframework
drf-nested-routers
drf-extensions

# librares
easy_thumbnails
whitenoise


django-allauth
django-imagekit
django-easy-select2
django-reversion==1.10.2

# short_url
# pyqrcode

# gunicorn
# psycopg2
# meinheld
# gevent
# greenlet==0.4.9

# Redis support
django-redis-cache
django-redis
redis
click
plan
django-rest-swagger
django-cors-headers
jpush
jinja2

django-suit-dashboard
short_url
#psutil

django-daterange-filter
django_celery_results
django_celery_beat
qrcode

<<<<<<< HEAD
rongcloud.zip
=======
# rongcloud.zip
>>>>>>> a2ab19053c4ae08232bdae1e407b9e9d05f7686c

drf-url-filters
django-url-filter

flower
#pandas
django-jsonfield
django-constance
honcho
arrow
raven